Refactor domain::Settlement
