+++ crates/autopilot/src/run_loop.rs
@@ -152,7 +152,7 @@ impl RunLoop {
             self.persistence
                 .store_order_events(order_uids, OrderEventLabel::Considered);
 
-            let winner = solution.account().into();
+            let winner = solution.solver().into();
             let winning_score = solution.score().get().0;
             let reference_score = solutions
                 .iter()
@@ -161,7 +161,7 @@ impl RunLoop {
                 .unwrap_or_default();
             let participants = solutions
                 .iter()
-                .map(|participant| participant.solution.account().into())
+                .map(|participant| participant.solution.solver().into())
                 .collect::<HashSet<_>>();
 
             let mut prices = BTreeMap::new();
@@ -221,7 +221,7 @@ impl RunLoop {
                         let is_winner = solutions.len() - index == 1;
                         let mut settlement = SolverSettlement {
                             solver: participant.driver.name.clone(),
-                            solver_address: participant.solution.account().0,
+                            solver_address: participant.solution.solver().0,
                             score: Some(Score::Solver(participant.solution.score().get().0)),
                             ranking: solutions.len() - index,
                             orders: participant
@@ -490,10 +490,10 @@ impl RunLoop {
             return auction;
         };
 
-        let tx_receipt = self.eth.transaction_receipt(in_flight.tx_hash.into()).await;
+        let transaction = self.eth.transaction(in_flight.tx_hash.into()).await;
 
-        let prev_settlement_block = match tx_receipt {
-            Ok(receipt) => receipt.block,
+        let prev_settlement_block = match transaction {
+            Ok(transaction) => transaction.block,
             // Could not find the block of the previous settlement, let's be
             // conservative and assume all orders are still in-flight.
             _ => u64::MAX.into(),
